Add checkpointing backends and configuration
Checkpointing can be enabled using different storages. Currently Azure Blob storage class is implemented as an example. Cassandra backend skeleton is included as a starting point.
Checkpointing configuration is exposed via the Configuration class. In particular the namespace for the data to be stored is exposed also via a `checkpointNamespace` method in the CheckpointBackend trait.

To create a new backend simply implement the read and write methods in CheckpointBackend trait.

To select which backend to use, set the `backendType` in the configuration. [the value is not used yet]