Keep alive stream. Changes to the public API.
To keep the stream alive the stream returns a keep alive message when there is nothing in the stream. This message is injected by the source and filtered out before reaching the sink.

The public API has been split, the methods for single partitions and entire hub are separate in two classes.