Fix regression on checkpointing logic
The "scheduleCheckpoint" boolean parameter, in combination with the cached actors, was causing the "write" to never be scheduled. The boolean parameter was introduced not to schedule the write when the caller was just trying to read the offset. The fix covers that intention by scheduling the "write" only after a caller sends a new offset to the checkpointing actor.