Update message specs, rename “MessageType” to “MessageSchema”, and “created” to “received”
The “created” message property is still present, but is meant to contain the time with the Device clock, as opposed to the time of the IoT Hub. The two timestamps can differ when a message is queued in the device, or when a device clock is not in sync with the time of IoT Hub.