Fix: remove CP parallelism to avoid contentions on table creation.
The recently added parallelism used when writing checkpoints to the storage, was leading to several contentions at the time of creating the checkpointing table. Cassandra (with the standard configuration) doesn’t seem to like the concurrent table creations, and was crashing with inconsistent state, probably due to the un-resolvable contentions. Given that offsets are not meant to be stored for every message, parallelism doesn’t seem to be a strong requirement. We can re-evaluate this decision when testing with 100+ partitions, in case of CP becoming a bottleneck.

Other changes:
* improve docs and code comments
* hide some internal methods, not required to be publicly accessible
* some code cleanup